requirements.txt
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
decorator==4.4.2
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
firebase-admin==5.0.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imap-tools==0.35.0
iniconfig==1.1.1
instabot==0.117.0
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
moviepy==1.0.3
msgpack==1.0.2
nltk==3.6.2
numpy==1.19.2
packaging==20.4
pandas==1.1.4
Pillow==8.0.0
pluggy==0.13.1
praw==7.1.0
prawcore==1.5.0
proglog==0.1.9
proto-plus==1.18.1
protobuf==3.17.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pymongo==3.11.3
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.12.0
rsa==4.7.2
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
toml==0.10.1
tqdm==4.50.2
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
